\name{InterSIM-package}
\alias{InterSIM-package}
\docType{package}
\title{
Simulation of inter-related genomic datasets 
}
\description{
InterSIM is an R package that generates three inter-related data set with realistic inter- and intra- relationships based on the DNA methylation, mRNA expression and protein expression from the TCGA ovarian cancer study.
}
\details{
\tabular{ll}{
Package: \tab InterSIM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-23\cr
License: \tab GPL>=2\cr
}
}
\author{
Prabhakar Chalise, Rama Raghavan, Brooke Fridley
Maintainer: Rama Raghavan
}
